requirements.txt
python-dotenv
pandas
numpy
pymongo[srv] ==3.6
pymongo
certifi
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart

# -e .